[REFACTOR] mv deps, crud_layer, crud_user to legacy
