+++ app/api/src/endpoints/v1/organizations.py
@@ -6,7 +6,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from src import crud, schemas
 from src.db import models
-from src.endpoints import deps
+from src.endpoints.legacy import deps
 from src.schemas.organization import request_examples
 
 router = APIRouter()
