+++ app/api/src/endpoints/v1/study_area.py
@@ -6,7 +6,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from src import crud
 from src.db import models
-from src.endpoints import deps
+from src.endpoints.legacy import deps
 from src.resources.enums import LayerGroupsEnum
 from src.schemas.study_area import pydantify_config
